# BENCH: peakmem benchmarks for heatmap